requirements.txt
absl-py==0.12.0
albumentations==0.5.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
greenlet==1.0.0
grpcio==1.36.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.7.2
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-core==4.7.1
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
neovim==0.3.1
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prettyprint==0.1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pynvim==0.4.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
tifffile==2021.3.5
toml==0.10.2
torch==1.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
